% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{solver}
\alias{solver}
\title{It uses basic functions to calculate sum, differece, product, 
             quoient, mean, exponential and logs}
\usage{
solver(x, y)
}
\arguments{
\item{x}{first numeric variable}

\item{y}{second numeric variable}
}
\description{
It uses basic functions to calculate sum, differece, product, 
             quoient, mean, exponential and logs
}
